Streamline player state management
Place all player state management into the playerservice object.  This
includes handling the local data store etc.

Still need to get the "relogin" code and the actual "playerChoice" modal
working, but overall there is a `playerService.get` method that will
return a promise for the current player, or if not, will pull down a
modal to allow the person to actually create a player.