Two major bugfixes
Add cleanup code to remove game-level listeners after when joining a new
game.  This fixed a number of bugs that would occur when players would
leave/rejoin game.  The tests weren't seeing this, because all of them
actually disconnected the sockets; update the test code so that the
sockets are not disconnected, to help test this.

The second big set of bugs had to do with how the transaction code was
being used: the transactions were only being applied to the central
portion of the event listerners, but not the validators.  Now the
transactions are applied to the entire sequence of events, so that the
validation can't be circumvented with a timing attack.

Also applied transactions to a few more important events (e.g. when
players leave the game).